pyproject.toml
[project]
name = "pydantic-validator"
version = "0.1.0"
description = ""
authors = [
    {name = "Wallace Camargo",email = "wallacecpdg@gmail.com"}
]
license = {text = "beerware"}
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "streamlit (==1.46.0)",
    "boto3 (==1.40.6)",
    "python-dotenv (==1.1.1)",
    "pydantic (==2.11.7)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
